Breaking: Group types with utils implementation
Fix #2392
