+++ packages/utils/src/config/index.ts
@@ -1,3 +1,4 @@
 export { default as normalizeHints } from './normalize-hints';
 export * from './to-absolute-paths';
 export * from './get-hints-from-configuration';
+export * from './types';
