+++ packages/utils/src/packages/enums.ts
@@ -1,3 +1,10 @@
+export enum ResourceErrorStatus {
+    DependencyError = 'DependencyError',
+    NotCompatible = 'NotCompatible',
+    NotFound = 'NotFound',
+    Unknown = 'Unknown'
+}
+
 /** The type of resource */
 export enum ResourceType {
     configuration = 'configuration',
