+++ packages/utils/src/packages/index.ts
@@ -1,3 +1,4 @@
+export * from './enums';
 export * from './find-node-modules-root';
 export * from './find-package-root';
 export * from './is-official';
@@ -7,3 +8,4 @@ export * from './has-multiple-resources';
 export * from './is-full-package-name';
 export * from './load-hint-package';
 export * from './require-package';
+export * from './resource-error';
