+++ packages/utils/src/packages/load-resource.ts
@@ -8,14 +8,13 @@ import { cwd, loadJSONFile } from '@hint/utils-fs';
 import { debug as d } from '@hint/utils-debug';
 import { normalizeIncludes } from '@hint/utils-string';
 import { isFullPackageName } from './is-full-package-name';
-import { ResourceType } from '../types/resource-type';
 import { loadPackage } from './load-package';
 import { loadHintPackage } from './load-hint-package';
 import { requirePackage } from './require-package';
 import { hasMultipleResources } from './has-multiple-resources';
 import { toAbsolutePaths } from '../config/to-absolute-paths';
-import { ResourceErrorStatus } from '../types/resource-error-status';
-import { ResourceError } from '../types/resource-error';
+import { ResourceErrorStatus, ResourceType } from './enums';
+import { ResourceError } from './resource-error';
 
 const debug: debug.IDebugger = d(__filename);
