+++ packages/utils/src/types/index.ts
@@ -1,5 +1,1 @@
-export * from './config';
 export * from './npm';
-export * from './resource-error';
-export * from './resource-error-status';
-export * from './resource-type';
